+++ controller/actionUpdateSupplier.php
@@ -27,116 +27,127 @@
         $changeGradeOrReference = "";
         $changeDocumentInfo = "";
 
-        if(!empty($manufacture)){
-            if($manufacture != $checkValue[0]['manufacture']){
-                $changeManufacture = ", manufacture";
-            }else{
-                $changeManufacture = "";
-            }
-        }else{
-            $manufacture = "-";
-        }
+        // Cek Apakah data Supplier tersedia
+        if($conn->query("SELECT * FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll()){
 
-        if(!empty($originCountry)){
-            if($originCountry != $checkValue[0]['originCountry']){
-                $changeOriginCountry = ", Origin Country";
-            }else{
-                $changeOriginCountry = "";
-            }
-        }else{
-            $originCountry = "-";
-        }
-
-        if(!empty($leadTime)){
-            if($leadTime != $checkValue[0]['leadTime']){
-                $changeLeadTime = ", Lead Time";
-            }else{
-                $changeLeadTime = "";
-            }
-        }else{
-            $leadTime = "-";
-        }
-
-        if(!empty($catalogOrCasNumber)){
-            if($catalogOrCasNumber != $checkValue[0]['catalogOrCasNumber']){
-                $changeCatalogOrCasNumber = ", Catalog Or Cas Number";
-            }else{
-                $changeCatalogOrCasNumber = "";
-            }
-        }else{
-            $catalogOrCasNumber = "-";
-        }
-
-        if(!empty($gradeOrReference)){
-            if($gradeOrReference != $checkValue[0]['gradeOrReference']){
-                $changeGradeOrReference = ", Grade Or Reference";
-            }else{
-                $changeGradeOrReference = "";
-            }
-        }else{
-            $gradeOrReference = "-";
-        }
-
-        if(!empty($documentInfo)){
-            if($documentInfo != $checkValue[0]['documentInfo']){
-                $changeDocumentInfo = ", Document Info";
-            }else{
-                $changeDocumentInfo = "";
-            }
-        }else{
-            $documentInfo = "-";
-        }
-
-        if(!empty($supplier)){
-            if($supplier != $checkValue[0]['supplier']){
-                $changeSupplier = ", supplier";
-            }else{
-                $changeSupplier = "";
-            }
-
-
-            if($conn->query("SELECT * FROM TB_MasterVendor WHERE vendorName = '".$supplier."'")->fetchAll()){
-                $sql = "UPDATE TB_Supplier SET supplier = ?, manufacture = ?, originCountry = ?, leadTime = ?, catalogOrCasNumber = ?, gradeOrReference = ?, documentInfo = ? WHERE id = ?";
-                $query = $conn->prepare($sql);
-                $update = $query->execute(array($supplier, $manufacture, $originCountry, $leadTime, $catalogOrCasNumber, $gradeOrReference, $documentInfo, $idSupplier));
-                
-                // Send Notifikasi
-                if($update == true){
-                    $message = "mengedit data".$changeSupplier.$changeManufacture.$changeOriginCountry.$changeLeadTime.$changeCatalogOrCasNumber.$changeGradeOrReference.$changeDocumentInfo.". Pada supplier : ";
-                    $dataSupplier = $conn->query("SELECT supplier, idMaterial FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll();
-                    $response = sendNotification("Supplier berhasil diedit!!", $dataSupplier[0]['supplier'], $message, NULL, $dataSupplier[0]['idMaterial'], $idSupplier);
+            if(!empty($manufacture)){
+                if($manufacture != $checkValue[0]['manufacture']){
+                    $changeManufacture = ", manufacture";
+                }else{
+                    $changeManufacture = "";
                 }
             }else{
-                // Memasukan data vendor ke database
-                $sqlAddVendor = "INSERT INTO TB_MasterVendor (vendorName) 
-                VALUES (?)";
-                $paramsAddVendor = array(
-                    $supplier
-                );
-                $queryAddVendor = $conn->prepare($sqlAddVendor);
-                $insertAddVendor =  $queryAddVendor->execute($paramsAddVendor);
+                $manufacture = "-";
+            }
 
-                if($insertAddVendor == true){
+            if(!empty($originCountry)){
+                if($originCountry != $checkValue[0]['originCountry']){
+                    $changeOriginCountry = ", Origin Country";
+                }else{
+                    $changeOriginCountry = "";
+                }
+            }else{
+                $originCountry = "-";
+            }
+
+            if(!empty($leadTime)){
+                if($leadTime != $checkValue[0]['leadTime']){
+                    $changeLeadTime = ", Lead Time";
+                }else{
+                    $changeLeadTime = "";
+                }
+            }else{
+                $leadTime = "-";
+            }
+
+            if(!empty($catalogOrCasNumber)){
+                if($catalogOrCasNumber != $checkValue[0]['catalogOrCasNumber']){
+                    $changeCatalogOrCasNumber = ", Catalog Or Cas Number";
+                }else{
+                    $changeCatalogOrCasNumber = "";
+                }
+            }else{
+                $catalogOrCasNumber = "-";
+            }
+
+            if(!empty($gradeOrReference)){
+                if($gradeOrReference != $checkValue[0]['gradeOrReference']){
+                    $changeGradeOrReference = ", Grade Or Reference";
+                }else{
+                    $changeGradeOrReference = "";
+                }
+            }else{
+                $gradeOrReference = "-";
+            }
+
+            if(!empty($documentInfo)){
+                if($documentInfo != $checkValue[0]['documentInfo']){
+                    $changeDocumentInfo = ", Document Info";
+                }else{
+                    $changeDocumentInfo = "";
+                }
+            }else{
+                $documentInfo = "-";
+            }
+
+            if(!empty($supplier)){
+                if($supplier != $checkValue[0]['supplier']){
+                    $changeSupplier = ", supplier";
+                }else{
+                    $changeSupplier = "";
+                }
+
+
+                if($conn->query("SELECT * FROM TB_MasterVendor WHERE vendorName = '".$supplier."'")->fetchAll()){
                     $sql = "UPDATE TB_Supplier SET supplier = ?, manufacture = ?, originCountry = ?, leadTime = ?, catalogOrCasNumber = ?, gradeOrReference = ?, documentInfo = ? WHERE id = ?";
                     $query = $conn->prepare($sql);
                     $update = $query->execute(array($supplier, $manufacture, $originCountry, $leadTime, $catalogOrCasNumber, $gradeOrReference, $documentInfo, $idSupplier));
-                
+                    
                     // Send Notifikasi
                     if($update == true){
                         $message = "mengedit data".$changeSupplier.$changeManufacture.$changeOriginCountry.$changeLeadTime.$changeCatalogOrCasNumber.$changeGradeOrReference.$changeDocumentInfo.". Pada supplier : ";
                         $dataSupplier = $conn->query("SELECT supplier, idMaterial FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll();
                         $response = sendNotification("Supplier berhasil diedit!!", $dataSupplier[0]['supplier'], $message, NULL, $dataSupplier[0]['idMaterial'], $idSupplier);
                     }
+                }else{
+                    // Memasukan data vendor ke database
+                    $sqlAddVendor = "INSERT INTO TB_MasterVendor (vendorName) 
+                    VALUES (?)";
+                    $paramsAddVendor = array(
+                        $supplier
+                    );
+                    $queryAddVendor = $conn->prepare($sqlAddVendor);
+                    $insertAddVendor =  $queryAddVendor->execute($paramsAddVendor);
+
+                    if($insertAddVendor == true){
+                        $sql = "UPDATE TB_Supplier SET supplier = ?, manufacture = ?, originCountry = ?, leadTime = ?, catalogOrCasNumber = ?, gradeOrReference = ?, documentInfo = ? WHERE id = ?";
+                        $query = $conn->prepare($sql);
+                        $update = $query->execute(array($supplier, $manufacture, $originCountry, $leadTime, $catalogOrCasNumber, $gradeOrReference, $documentInfo, $idSupplier));
+                    
+                        // Send Notifikasi
+                        if($update == true){
+                            $message = "mengedit data".$changeSupplier.$changeManufacture.$changeOriginCountry.$changeLeadTime.$changeCatalogOrCasNumber.$changeGradeOrReference.$changeDocumentInfo.". Pada supplier : ";
+                            $dataSupplier = $conn->query("SELECT supplier, idMaterial FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll();
+                            $response = sendNotification("Supplier berhasil diedit!!", $dataSupplier[0]['supplier'], $message, NULL, $dataSupplier[0]['idMaterial'], $idSupplier);
+                        }
+                    }
                 }
+            }else{
+                $response = array(
+                    "status" => 1,
+                    "message" => "Data supplier wajib di isi!!"
+                );
             }
+
         }else{
             $response = array(
                 "status" => 1,
-                "message" => "Data supplier wajib di isi!!"
+                "message" => "Data supplier tidak ditemukan", 
             );
         }
 
         echo json_encode($response);
+        exit();
     }
 
     if(isset($_POST['addDetail'])){
@@ -149,44 +160,66 @@
 
         $priceDetail = $hardCash.$price.$quantity;
 
-        $sql = "INSERT INTO TB_DetailSupplier (MoQ, UoM, price, idSupplier) 
-        VALUES (?,?,?,?)";
-        $params = array(
-            $MoQ,
-            $UoM,
-            $priceDetail,
-            $idSupplier,
-        );
-        $query = $conn->prepare($sql);
-        $insert = $query->execute($params);
+        // Cek Apakah data Supplier tersedia
+        if($conn->query("SELECT * FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll()){
 
-        // Send Notifikasi
-        if($insert == true){
-            $dataSupplier = $conn->query("SELECT supplier, idMaterial FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll();
-            $response = sendNotification("Berhasil memasukan detail supplier!!", $dataSupplier[0]['supplier'] , "menambahkan detail supplier : ", NULL, $dataSupplier[0]['idMaterial'], $idSupplier);
+            $sql = "INSERT INTO TB_DetailSupplier (MoQ, UoM, price, idSupplier) 
+            VALUES (?,?,?,?)";
+            $params = array(
+                $MoQ,
+                $UoM,
+                $priceDetail,
+                $idSupplier,
+            );
+            $query = $conn->prepare($sql);
+            $insert = $query->execute($params);
+
+            // Send Notifikasi
+            if($insert == true){
+                $dataSupplier = $conn->query("SELECT supplier, idMaterial FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll();
+                $response = sendNotification("Berhasil memasukan detail supplier!!", $dataSupplier[0]['supplier'] , "menambahkan detail supplier : ", NULL, $dataSupplier[0]['idMaterial'], $idSupplier);
+            }
+
+        }else{
+            $response = array(
+                "status" => 1,
+                "message" => "Data supplier tidak ditemukan", 
+            );
         }
 
     
         echo json_encode($response);
+        exit();
     }
 
     // Kondisi saat menghapus detail Supplier
     if(($_REQUEST['actionType'] == 'delete') && !empty($_GET['idDetailSupplier'])){
-        $idDetailMaterial = $_GET['idDetailSupplier'];
+        $idDetailSupplier = $_GET['idDetailSupplier'];
         $idSupplier = $_GET['idSupplier'];
 
-        //Delete data from SQL server 
-        $sql = "DELETE FROM TB_DetailSupplier WHERE idDetailSupplier = ?"; 
-        $query = $conn->prepare($sql); 
-        $delete = $query->execute(array($idDetailMaterial));
-        
-        // Send Notifikasi
-        if($delete == true){
-            $dataSupplier = $conn->query("SELECT supplier, idMaterial FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll();
-            $response = sendNotification("Berhasil menghapus detail supplier!!", $dataSupplier[0]['supplier'], "menghapus detail supplier : ", NULL, $dataSupplier[0]['idMaterial'], $idSupplier);
+        // Cek Apakah data Supplier tersedia
+        if($conn->query("SELECT * FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll()){
+
+            //Delete data from SQL server 
+            $sql = "DELETE FROM TB_DetailSupplier WHERE idDetailSupplier = ?"; 
+            $query = $conn->prepare($sql); 
+            $delete = $query->execute(array($idDetailSupplier));
+            
+            // Send Notifikasi
+            if($delete == true){
+                $dataSupplier = $conn->query("SELECT supplier, idMaterial FROM TB_Supplier WHERE id = ".$idSupplier)->fetchAll();
+                $response = sendNotification("Berhasil menghapus detail supplier!!", $dataSupplier[0]['supplier'], "menghapus detail supplier : ", NULL, $dataSupplier[0]['idMaterial'], $idSupplier);
+            }
+
+        }else{
+            $response = array(
+                "status" => 1,
+                "message" => "Data supplier tidak ditemukan", 
+            );
         }
-        // $response = $_GET['idSupplier'];
+
         echo json_encode($response);
+        exit();
     }
 
     // Function For Send Nofitication
